amends to make background-switcher work when using pageLevelProgres to navigate to a block that has no background set (see issue #1)
also some re-factoring of code - particularly in the 'find background' code as was using forEach which doesn't actually support exiting via return, so switched to using _.find which does. also created an onPageViewReady method to aggregate all the functions that need to be run when that event is called - instead of having them all attached separately. not sure if performance is any better but looks more readable to me!